imp: initial attempt

+++ contracts/ICS20Transfer.sol
@@ -141,7 +141,7 @@ contract ICS20Transfer is
         external
         whenNotPaused
         nonReentrant
-        returns (uint32)
+        returns (uint64)
     {
         require(msg_.amount > 0, IICS20Errors.ICS20InvalidAmount(0));
         // transfer the tokens to us (requires the allowance to be set)
@@ -161,7 +161,7 @@ contract ICS20Transfer is
         external
         whenNotPaused
         nonReentrant
-        returns (uint32)
+        returns (uint64)
     {
         require(msg_.amount > 0, IICS20Errors.ICS20InvalidAmount(0));
         require(
@@ -190,7 +190,7 @@ contract ICS20Transfer is
         address escrow
     )
         private
-        returns (uint32)
+        returns (uint64)
     {
         string memory fullDenomPath = _getICS20TransferStorage()._ibcERC20Denoms[msg_.denom];
         if (bytes(fullDenomPath).length == 0) {
