+++ test/solidity-ibc/ICS20TransferTest.t.sol
@@ -85,7 +85,7 @@ contract ICS20TransferTest is Test, DeployPermit2, PermitSignature {
 
         vm.mockCall(address(this), abi.encodeWithSelector(IICS26Router.sendPacket.selector), abi.encode(uint32(42)));
         vm.prank(sender);
-        uint32 sequence = ics20Transfer.sendTransfer(msgSendTransfer);
+        uint64 sequence = ics20Transfer.sendTransfer(msgSendTransfer);
         assertEq(sequence, 42);
     }
